refactor: extract CLI commands into modular structure
- Split cli.py (3556 lines) into focused command modules
- Create cli/commands/ directory with separate files for each command group:
  - config.py: Configuration management commands
  - database.py: Database operations and management
  - script.py: Screenplay parsing and management
  - scene.py: Scene analysis and management
  - search.py: Search and query operations
  - dev.py: Development and debugging utilities
  - bible.py: Script Bible and continuity management
  - server.py: MCP server operations
  - mentor.py: Screenplay analysis mentors
- Add cli/main.py as unified entry point importing all command groups
- Add __main__.py to support 'python -m scriptrag' execution
- Fix mypy duplicate module error by removing old cli.py
- Fix mentor.py method call from save_analysis_result to store_mentor_result
- Fix script.py error handling for dict-type errors field
- Maintain full CLI functionality and backward compatibility
- Reduce main CLI complexity while preserving all features
