+++ src/scriptrag/__main__.py
@@ -1,0 +1,6 @@
+"""Main entry point for ScriptRAG when run as a module."""
+
+from .cli import app
+
+if __name__ == "__main__":
+    app()
