+++ src/scriptrag/cli/__init__.py
@@ -1,0 +1,5 @@
+"""ScriptRAG CLI package."""
+
+from .main import app
+
+__all__ = ["app"]
