+++ src/scriptrag/cli/commands/__init__.py
@@ -1,0 +1,1 @@
+"""ScriptRAG CLI commands package."""
